Move theme into popup code
